[Fix #17] Don't add non-title entries to imenu (#18)
Fixes #17 by checking that 2-line titles have the same number of
under characters as the title text and doesn't look like the start
of an attribute block.

Also check when (adoc-title-descriptor) returns nil, which it will at
times when the regexes match but more stringent tests fail, a nil list
will be inserted into the imenu list, which it doesn't like much, so
this condition is checked for as well.

This also add a strict mode to adoc-title-descriptor so it can work for some of
the sloppy editing scenarios needed. Also make it conform to asciidoc
standard of allowing as many as two character difference in two line
titles, in strict mode.